Falling Workpieces Simulation and Pose Finder
This is specifically for the items related to the past work done on simulating three dimensional objects falling on a flat surface and the evaluation of their resting poses. The folder with the MATLAB code is the original written by ToKomatch, the PoseFinder is the version made in Python with the intent on further work continuing in Python with the intent of a better interface between the simulating part and the pose finding part.